Name the shared executor code for what it is
None of this changes behaviour; it stops backend-neutral code being named
after one backend.

run_execution.rs runs the planner for OpenFold and ESMFold alike -- its
own comment said so -- and run_artifacts.rs registers a run's output
directories whichever backend produced them. Both were openfold_*.

The planner itself moves to model_runners/plan.rs as plan_command. It is
declared by each backend's invocation profile in JSON and shared by both,
and it living in the OpenFold runner is most of why that file was ten
times the size of the ESMFold one. Same for the checks that validate a
planned command rather than a model's inputs (base_command_checks,
detect_gpu, gpu_check, input_id_check, output_dir_check): they are now in
core/preflight.rs, where esmfold.rs can import them without reaching into
the other backend.

openfold.rs is 1786 lines down to 1254, and what is left is OpenFold's
own MSA, template and alignment checks. esmfold.rs is unchanged at 167,
so the honest ratio is about 7:1 and it is real.

Pure moves: cargo test passes with no test body edited, only the import
lines that name where the functions now live.
