+++ cli/src/core/model_runners/mod.rs
@@ -1,2 +1,3 @@
 pub mod esmfold;
 pub mod openfold;
+pub mod plan;
