+++ cli/src/core/model_runners/openfold.rs
@@ -1,87 +1,18 @@
-use std::{
-    collections::BTreeMap,
-    path::{Path, PathBuf},
-};
+use std::path::{Path, PathBuf};
 
 use sea_orm::DbErr;
 use serde_json::Value;
 
+use super::plan::{optional_bool, optional_string, parse_object};
 use crate::core::{
     commands::CommandSpec,
-    entities::{execution_targets, model_backends, model_invocation_profiles, runs},
+    entities::{model_invocation_profiles, runs},
     output_locations::resolve_output_location,
-    preflight::{PreflightCheck, PreflightReport},
+    preflight::{
+        PreflightCheck, PreflightReport, base_command_checks, detect_gpu, gpu_check,
+        input_id_check, output_dir_check,
+    },
 };
-
-pub fn plan_openfold_command(
-    model_backend: &model_backends::Model,
-    execution_target: &execution_targets::Model,
-    invocation_profile: &model_invocation_profiles::Model,
-    run: &runs::Model,
-) -> Result<CommandSpec, DbErr> {
-    validate_entity_consistency(invocation_profile)?;
-
-    let config = parse_object(
-        "model invocation profile config_json",
-        &invocation_profile.config_json,
-    )?;
-    let model_schema = parse_object(
-        "model backend parameter_schema_json",
-        &model_backend.parameter_schema_json,
-    )?;
-    let available_resources = parse_object(
-        "execution target available_resources_json",
-        &execution_target.available_resources_json,
-    )?;
-    let model_parameters = parse_object("run model_parameters_json", &run.model_parameters_json)?;
-    let execution_parameters = parse_object(
-        "run execution_parameters_json",
-        &run.execution_parameters_json,
-    )?;
-    validate_execution_parameters_against_available_resources(
-        &available_resources,
-        &execution_parameters,
-    )?;
-
-    let program = required_string(&config, "program")?;
-    let script = required_string(&config, "script")?;
-    let current_dir = optional_string(&config, "working_dir").map(PathBuf::from);
-    let env = parse_env(&config)?;
-    let mut args = vec!["-u".into(), script];
-
-    append_model_schema_args(
-        &mut args,
-        &model_schema,
-        &model_parameters,
-        &execution_parameters,
-        &config,
-        invocation_profile,
-        run,
-    )?;
-
-    append_available_resources_args(&mut args, &available_resources, &execution_parameters);
-
-    if let Some(residue_idx) = optional_i64(&execution_parameters, "residue_idx") {
-        args.extend(["--triangle_residue_idx".into(), residue_idx.to_string()]);
-    }
-
-    if optional_bool(&execution_parameters, "use_precomputed_alignments").unwrap_or(false) {
-        let alignment_dir = required_string(&execution_parameters, "alignment_dir")?;
-        args.extend(["--use_precomputed_alignments".into(), alignment_dir]);
-    }
-
-    // Intentionally do not emit model_preset. The OpenFold script used by this
-    // repository currently exposes --config_preset, and model_preset is not part
-    // of the MVP OpenFold parameter schema.
-
-    Ok(CommandSpec {
-        program,
-        args,
-        current_dir,
-        env,
-        stream: false,
-    })
-}
 
 pub fn preflight_openfold(
     command: &CommandSpec,
@@ -115,108 +46,6 @@ pub fn preflight_openfold(
     Ok(PreflightReport::new(checks))
 }
 
-/// Mirrors the fold's own `nvidia-smi --query-gpu=name --format=csv,noheader` probe.
-pub fn detect_gpu() -> Option<String> {
-    let output = std::process::Command::new("nvidia-smi")
-        .args(["--query-gpu=name", "--format=csv,noheader"])
-        .output()
-        .ok()?;
-    let name = String::from_utf8_lossy(&output.stdout)
-        .lines()
-        .next()?
-        .trim()
-        .to_owned();
-    (output.status.success() && !name.is_empty()).then_some(name)
-}
-
-pub fn gpu_check(detected: Option<&str>) -> PreflightCheck {
-    match detected {
-        Some(name) => PreflightCheck::passed("gpu", format!("GPU visible: {name}")),
-        None => PreflightCheck::warning("gpu", "no GPU visible; the run will fall back to CPU"),
-    }
-}
-
-/// Program/script-argument/working-directory/script-file checks shared by every local_subprocess
-/// backend's preflight (OpenFold, ESMFold): they validate the planned command itself, independent
-/// of a backend's inputs. Callers push these after `gpu_check` and before their own input checks.
-pub(super) fn base_command_checks(command: &CommandSpec) -> Vec<PreflightCheck> {
-    let program = if command.program.trim().is_empty() {
-        PreflightCheck::failed("program configured", "command program is empty")
-    } else {
-        PreflightCheck::passed(
-            "program configured",
-            format!("program '{}' is configured", command.program),
-        )
-    };
-
-    let script = script_argument(command);
-    let script_arg = match script {
-        Some(script) => PreflightCheck::passed(
-            "script argument configured",
-            format!("script argument '{script}' follows -u"),
-        ),
-        None => PreflightCheck::failed(
-            "script argument configured",
-            "command args must include a script argument after -u",
-        ),
-    };
-
-    let working_dir = match &command.current_dir {
-        Some(current_dir) if current_dir.is_dir() => PreflightCheck::passed(
-            "working directory",
-            format!("working directory '{}' exists", current_dir.display()),
-        ),
-        Some(current_dir) => PreflightCheck::failed(
-            "working directory",
-            format!(
-                "working directory '{}' does not exist or is not a directory",
-                current_dir.display()
-            ),
-        ),
-        None => PreflightCheck::warning(
-            "working directory",
-            "no working directory is configured; script resolution may depend on the caller",
-        ),
-    };
-
-    let script_file = match (script, &command.current_dir) {
-        (Some(script), _) if Path::new(script).is_absolute() => {
-            path_exists_check("script file", Path::new(script))
-        }
-        (Some(script), Some(current_dir)) => {
-            path_exists_check("script file", &current_dir.join(script))
-        }
-        (Some(script), None) => PreflightCheck::warning(
-            "script file",
-            format!("relative script '{script}' cannot be resolved without a working directory"),
-        ),
-        (None, _) => PreflightCheck::failed(
-            "script file",
-            "script path is unavailable because the -u script argument is missing",
-        ),
-    };
-
-    vec![program, script_arg, working_dir, script_file]
-}
-
-fn script_argument(command: &CommandSpec) -> Option<&str> {
-    command
-        .args
-        .iter()
-        .position(|arg| arg == "-u")
-        .and_then(|index| command.args.get(index + 1))
-        .map(String::as_str)
-        .filter(|script| !script.is_empty())
-}
-
-pub(super) fn path_exists_check(name: &str, path: &Path) -> PreflightCheck {
-    if path.exists() {
-        PreflightCheck::passed(name, format!("'{}' exists", path.display()))
-    } else {
-        PreflightCheck::failed(name, format!("'{}' does not exist", path.display()))
-    }
-}
-
 fn required_directory_check(parameters: &Value, field_name: &str) -> PreflightCheck {
     let Some(path) = optional_string(parameters, field_name).filter(|path| !path.is_empty()) else {
         return PreflightCheck::failed(field_name, format!("{field_name} is missing"));
@@ -228,17 +57,6 @@ fn required_directory_check(parameters: &Value, field_name: &str) -> PreflightCh
         PreflightCheck::failed(
             field_name,
             format!("'{path}' does not exist or is not a directory"),
-        )
-    }
-}
-
-pub(super) fn input_id_check(input_id: &str) -> PreflightCheck {
-    if input_id.trim().is_empty() {
-        PreflightCheck::failed("input_id", "run input_id is missing or empty")
-    } else {
-        PreflightCheck::passed(
-            "input_id",
-            format!("run input_id '{input_id}' is configured"),
         )
     }
 }
@@ -403,354 +221,6 @@ fn precomputed_alignment_key_check(parameters: &Value, input_id: &str) -> Prefli
     }
 }
 
-pub(super) fn output_dir_check(output_path: &Path) -> PreflightCheck {
-    if output_path.exists() {
-        return if output_path.is_dir() {
-            PreflightCheck::passed(
-                "output_dir parent",
-                format!("'{}' already exists", output_path.display()),
-            )
-        } else {
-            PreflightCheck::failed(
-                "output_dir parent",
-                format!("'{}' exists but is not a directory", output_path.display()),
-            )
-        };
-    }
-
-    let parent = output_path
-        .parent()
-        .filter(|path| !path.as_os_str().is_empty())
-        .unwrap_or_else(|| Path::new("."));
-    if parent.is_dir() {
-        PreflightCheck::passed(
-            "output_dir parent",
-            format!("parent '{}' exists", parent.display()),
-        )
-    } else {
-        PreflightCheck::failed(
-            "output_dir parent",
-            format!(
-                "parent '{}' does not exist or is not a directory",
-                parent.display()
-            ),
-        )
-    }
-}
-
-fn validate_entity_consistency(
-    invocation_profile: &model_invocation_profiles::Model,
-) -> Result<(), DbErr> {
-    if invocation_profile.invocation_kind != "local_subprocess" {
-        return Err(DbErr::Custom(format!(
-            "OpenFold planner only supports local_subprocess invocation profiles, got '{}'",
-            invocation_profile.invocation_kind
-        )));
-    }
-
-    Ok(())
-}
-
-fn append_model_schema_args(
-    args: &mut Vec<String>,
-    model_schema: &Value,
-    model_parameters: &Value,
-    execution_parameters: &Value,
-    invocation_config: &Value,
-    invocation_profile: &model_invocation_profiles::Model,
-    run: &runs::Model,
-) -> Result<(), DbErr> {
-    for (_name, declaration) in sorted_schema_declarations(model_schema, true) {
-        if optional_bool(declaration, "positional").unwrap_or(false) {
-            args.push(resolve_declared_value(
-                declaration,
-                model_parameters,
-                execution_parameters,
-                invocation_config,
-                invocation_profile,
-                run,
-            )?);
-        }
-    }
-
-    for (name, declaration) in sorted_schema_declarations(model_schema, false) {
-        if optional_bool(declaration, "positional").unwrap_or(false) {
-            continue;
-        }
-
-        let Some(cli_flag) = optional_string(declaration, "cli_flag") else {
-            continue;
-        };
-
-        if optional_string(declaration, "type").as_deref() == Some("boolean") {
-            if optional_bool(model_parameters, name).unwrap_or(false) {
-                args.push(cli_flag);
-            }
-            continue;
-        }
-
-        if declaration.get("source").is_some() {
-            let value = resolve_declared_value(
-                declaration,
-                model_parameters,
-                execution_parameters,
-                invocation_config,
-                invocation_profile,
-                run,
-            )?;
-            args.extend([cli_flag, value]);
-            continue;
-        }
-
-        if let Some(value) = selected_or_default_string(model_parameters, declaration, name) {
-            args.extend([cli_flag, value]);
-        }
-    }
-
-    Ok(())
-}
-
-fn append_available_resources_args(
-    args: &mut Vec<String>,
-    available_resources: &Value,
-    execution_parameters: &Value,
-) {
-    for (name, declaration) in sorted_schema_declarations(available_resources, false) {
-        let Some(cli_flag) = optional_string(declaration, "cli_flag") else {
-            continue;
-        };
-
-        if let Some(value) = selected_or_default_string(execution_parameters, declaration, name) {
-            args.extend([cli_flag, value]);
-        }
-    }
-}
-
-fn sorted_schema_declarations(schema: &Value, position_first: bool) -> Vec<(&str, &Value)> {
-    let mut declarations = schema
-        .get("properties")
-        .and_then(Value::as_object)
-        .map(|properties| {
-            properties
-                .iter()
-                .map(|(name, declaration)| (name.as_str(), declaration))
-                .collect::<Vec<_>>()
-        })
-        .unwrap_or_default();
-
-    declarations.sort_by(|(left_name, left), (right_name, right)| {
-        let left_position = optional_i64(left, "position").unwrap_or(i64::MAX);
-        let right_position = optional_i64(right, "position").unwrap_or(i64::MAX);
-        if position_first {
-            left_position
-                .cmp(&right_position)
-                .then_with(|| left_name.cmp(right_name))
-        } else {
-            left_name.cmp(right_name)
-        }
-    });
-
-    declarations
-}
-
-fn resolve_declared_value(
-    declaration: &Value,
-    _model_parameters: &Value,
-    execution_parameters: &Value,
-    invocation_config: &Value,
-    invocation_profile: &model_invocation_profiles::Model,
-    run: &runs::Model,
-) -> Result<String, DbErr> {
-    let source = required_string(declaration, "source")?;
-    match source.as_str() {
-        "data_dir" => {
-            let data_dir = required_string(execution_parameters, "data_dir")?;
-            let relative_path = required_string(declaration, "relative_path")?;
-            Ok(data_path(&data_dir, &relative_path))
-        }
-        "execution_parameters" => {
-            let parameter_name = required_string(declaration, "parameter")?;
-            required_string(execution_parameters, &parameter_name)
-        }
-        "invocation_profile_config" => {
-            let parameter_name = required_string(declaration, "parameter")?;
-            let value =
-                required_invocation_profile_config_string(invocation_config, &parameter_name)?;
-            let mut path = PathBuf::from(value);
-            if let Some(relative_path) = optional_string(declaration, "relative_path") {
-                path.push(relative_path);
-            }
-            Ok(path.to_string_lossy().into_owned())
-        }
-        "run_output_workspace" => {
-            let workspace = resolve_output_location(invocation_profile, run)?;
-            let path = optional_string(declaration, "relative_path")
-                .map(|relative_path| workspace.join(relative_path))
-                .unwrap_or(workspace);
-            Ok(path.to_string_lossy().into_owned())
-        }
-        _ => Err(DbErr::Custom(format!(
-            "unsupported model parameter source '{source}'"
-        ))),
-    }
-}
-
-fn required_invocation_profile_config_string(
-    config: &Value,
-    parameter_name: &str,
-) -> Result<String, DbErr> {
-    let Some(value) = config.get(parameter_name) else {
-        return Err(DbErr::Custom(format!(
-            "invocation profile config '{parameter_name}' is required"
-        )));
-    };
-    let Some(value) = value.as_str().filter(|value| !value.trim().is_empty()) else {
-        return Err(DbErr::Custom(format!(
-            "invocation profile config '{parameter_name}' must be a non-empty string"
-        )));
-    };
-
-    Ok(value.to_owned())
-}
-
-fn selected_or_default_string(
-    parameters: &Value,
-    declaration: &Value,
-    field_name: &str,
-) -> Option<String> {
-    parameters
-        .get(field_name)
-        .or_else(|| declaration.get("default"))
-        .and_then(json_value_to_string)
-}
-
-fn json_value_to_string(value: &Value) -> Option<String> {
-    match value {
-        Value::String(value) => Some(value.clone()),
-        Value::Number(value) => Some(value.to_string()),
-        Value::Bool(value) => Some(value.to_string()),
-        _ => None,
-    }
-}
-
-fn validate_execution_parameters_against_available_resources(
-    available_resources: &Value,
-    execution_parameters: &Value,
-) -> Result<(), DbErr> {
-    let Some(properties) = available_resources
-        .get("properties")
-        .and_then(Value::as_object)
-    else {
-        return Ok(());
-    };
-
-    // A present field is validated against its raw JSON type, not skipped when the
-    // type is wrong: emission (json_value_to_string) stringifies numbers/bools, so a
-    // wrong-typed value would otherwise bypass the enum/range guard yet still be emitted.
-    if let Some(declaration) = properties.get("model_device")
-        && let Some(value) = execution_parameters.get("model_device")
-    {
-        let model_device = value
-            .as_str()
-            .ok_or_else(|| DbErr::Custom("model_device must be a string".into()))?;
-        if let Some(allowed_values) = declaration.get("enum").and_then(Value::as_array) {
-            let allowed = allowed_values
-                .iter()
-                .filter_map(Value::as_str)
-                .collect::<Vec<_>>();
-
-            if !allowed.contains(&model_device) {
-                return Err(DbErr::Custom(format!(
-                    "model_device '{model_device}' is not allowed by execution target available resources"
-                )));
-            }
-        }
-    }
-
-    if let Some(declaration) = properties.get("cpus")
-        && let Some(value) = execution_parameters.get("cpus")
-    {
-        let cpus = value
-            .as_i64()
-            .ok_or_else(|| DbErr::Custom("cpus must be an integer".into()))?;
-        if let Some(minimum) = optional_i64(declaration, "minimum")
-            && cpus < minimum
-        {
-            return Err(DbErr::Custom(format!(
-                "cpus {cpus} is below execution target resource minimum {minimum}"
-            )));
-        }
-
-        if let Some(maximum) = optional_i64(declaration, "maximum")
-            && cpus > maximum
-        {
-            return Err(DbErr::Custom(format!(
-                "cpus {cpus} exceeds execution target resource maximum {maximum}"
-            )));
-        }
-    }
-
-    Ok(())
-}
-
-fn parse_object(field_name: &str, raw: &str) -> Result<Value, DbErr> {
-    let value: Value = serde_json::from_str(raw)
-        .map_err(|error| DbErr::Custom(format!("{field_name} must be valid JSON: {error}")))?;
-
-    if !value.is_object() {
-        return Err(DbErr::Custom(format!("{field_name} must be a JSON object")));
-    }
-
-    Ok(value)
-}
-
-fn parse_env(config: &Value) -> Result<BTreeMap<String, String>, DbErr> {
-    let Some(env) = config.get("env") else {
-        return Ok(BTreeMap::new());
-    };
-
-    let Some(env_object) = env.as_object() else {
-        return Err(DbErr::Custom("config env must be a JSON object".into()));
-    };
-
-    let mut parsed = BTreeMap::new();
-    for (key, value) in env_object {
-        let Some(value) = value.as_str() else {
-            return Err(DbErr::Custom(format!(
-                "config env value for '{key}' must be a string"
-            )));
-        };
-        parsed.insert(key.clone(), value.to_owned());
-    }
-
-    Ok(parsed)
-}
-
-fn required_string(object: &Value, field_name: &str) -> Result<String, DbErr> {
-    optional_string(object, field_name)
-        .filter(|value| !value.is_empty())
-        .ok_or_else(|| DbErr::Custom(format!("{field_name} is required")))
-}
-
-fn optional_string(object: &Value, field_name: &str) -> Option<String> {
-    object
-        .get(field_name)
-        .and_then(Value::as_str)
-        .map(str::to_owned)
-}
-
-fn optional_bool(object: &Value, field_name: &str) -> Option<bool> {
-    object.get(field_name).and_then(Value::as_bool)
-}
-
-fn optional_i64(object: &Value, field_name: &str) -> Option<i64> {
-    object.get(field_name).and_then(Value::as_i64)
-}
-
-fn data_path(data_dir: &str, suffix: &str) -> String {
-    format!("{}/{}", data_dir.trim_end_matches(['/', '\\']), suffix)
-}
-
 #[cfg(test)]
 mod tests {
     use std::{env, fs, path::PathBuf};
@@ -766,10 +236,8 @@ mod tests {
         test_support::TestLayout,
     };
 
-    use super::{
-        plan_openfold_command, preflight_openfold as preflight_openfold_impl,
-        resolve_declared_value,
-    };
+    use super::preflight_openfold as preflight_openfold_impl;
+    use crate::core::model_runners::plan::{plan_command, resolve_declared_value};
 
     fn preflight_run(execution_parameters: serde_json::Value) -> runs::Model {
         preflight_run_with_input_id("1UBQ_1", execution_parameters)
@@ -996,7 +464,7 @@ mod tests {
 
     #[test]
     fn builds_basic_openfold_command_spec() {
-        let command = plan_openfold_command(
+        let command = plan_command(
             &model_backend(),
             &execution_target(),
             &invocation_profile(config()),
@@ -1035,7 +503,7 @@ mod tests {
         execution["output_dir"] = json!("/stale/output");
         execution["attn_map_dir"] = json!("/stale/attention");
 
-        let command = plan_openfold_command(
+        let command = plan_command(
             &model_backend(),
             &execution_target(),
             &invocation_profile(config()),
@@ -1056,7 +524,7 @@ mod tests {
 
     #[test]
     fn schema_declared_output_paths_require_profile_output_location() {
-        let error = plan_openfold_command(
+        let error = plan_command(
             &model_backend(),
             &execution_target(),
             &invocation_profile(
@@ -1129,7 +597,7 @@ mod tests {
         let mut execution = execution_parameters();
         execution["model_device"] = json!("cpu");
 
-        let command = plan_openfold_command(
+        let command = plan_command(
             &model_backend(),
             &execution_target(),
             &invocation_profile(config()),
@@ -1172,7 +640,7 @@ mod tests {
             let mut execution = execution_parameters();
             execution[field] = value;
 
-            let command = plan_openfold_command(
+            let command = plan_command(
                 &model_backend(),
                 &execution_target(),
                 &invocation_profile(config()),
@@ -1189,7 +657,7 @@ mod tests {
         let mut execution = execution_parameters();
         execution["model_device"] = json!("cuda:1");
 
-        let error = plan_openfold_command(
+        let error = plan_command(
             &model_backend(),
             &execution_target(),
             &invocation_profile(config()),
@@ -1209,7 +677,7 @@ mod tests {
         let mut execution = execution_parameters();
         execution["cpus"] = json!(15);
 
-        let error = plan_openfold_command(
+        let error = plan_command(
             &model_backend(),
             &execution_target(),
             &invocation_profile(config()),
@@ -1225,7 +693,7 @@ mod tests {
         let mut execution = execution_parameters();
         execution["cpus"] = json!("999");
 
-        let error = plan_openfold_command(
+        let error = plan_command(
             &model_backend(),
             &execution_target(),
             &invocation_profile(config()),
@@ -1238,7 +706,7 @@ mod tests {
 
     #[test]
     fn database_paths_are_generated_from_model_schema_declarations() {
-        let command = plan_openfold_command(
+        let command = plan_command(
             &model_backend(),
             &execution_target(),
             &invocation_profile(config()),
@@ -1263,7 +731,7 @@ mod tests {
         let mut execution = execution_parameters();
         execution["residue_idx"] = json!(7);
 
-        let command = plan_openfold_command(
+        let command = plan_command(
             &model_backend(),
             &execution_target(),
             &invocation_profile(config()),
@@ -1293,7 +761,7 @@ mod tests {
         execution["use_precomputed_alignments"] = json!(true);
         execution["alignment_dir"] = json!("/tmp/alignments");
 
-        let command = plan_openfold_command(
+        let command = plan_command(
             &model_backend(),
             &execution_target(),
             &invocation_profile(config()),
@@ -1314,7 +782,7 @@ mod tests {
         let mut execution = execution_parameters();
         execution["use_precomputed_alignments"] = json!(true);
 
-        let error = plan_openfold_command(
+        let error = plan_command(
             &model_backend(),
             &execution_target(),
             &invocation_profile(config()),
@@ -1327,7 +795,7 @@ mod tests {
 
     #[test]
     fn returns_clear_error_when_required_config_field_is_missing() {
-        let error = plan_openfold_command(
+        let error = plan_command(
             &model_backend(),
             &execution_target(),
             &invocation_profile(json!({"script": "run_pretrained_openfold.py"}).to_string()),
@@ -1340,7 +808,7 @@ mod tests {
 
     #[test]
     fn returns_clear_error_when_required_execution_parameter_is_missing() {
-        let error = plan_openfold_command(
+        let error = plan_command(
             &model_backend(),
             &execution_target(),
             &invocation_profile(config()),
@@ -1359,7 +827,7 @@ mod tests {
 
     #[test]
     fn returns_clear_error_when_schema_declared_fasta_dir_is_missing() {
-        let error = plan_openfold_command(
+        let error = plan_command(
             &model_backend(),
             &execution_target(),
             &invocation_profile(config()),
@@ -1379,7 +847,7 @@ mod tests {
 
     #[test]
     fn validates_env_is_string_to_string_object() {
-        let error = plan_openfold_command(
+        let error = plan_command(
             &model_backend(),
             &execution_target(),
             &invocation_profile(
@@ -1406,7 +874,7 @@ mod tests {
         let mut profile = invocation_profile(config());
         profile.invocation_kind = "docker".into();
 
-        let error = plan_openfold_command(
+        let error = plan_command(
             &model_backend(),
             &execution_target(),
             &profile,
