+++ cli/src/core/services/mod.rs
@@ -3,7 +3,7 @@ pub mod artifacts;
 pub mod execution_targets;
 pub mod model_backends;
 pub mod model_invocation_profiles;
-pub mod openfold_artifacts;
-pub mod openfold_execution;
+pub mod run_artifacts;
+pub mod run_execution;
 pub mod runs;
 mod validation;
