+++ docs/vizfold_gateway/architecture.md
@@ -22,7 +22,7 @@ This diagram describes the MVP data model for the Rust executor core. The goal i
 
 This model intentionally does not include model-target artifact constraint logic in the MVP. Artifact capabilities remain model-level, while actual produced outputs are recorded as `ARTIFACT` rows classified by the `ARTIFACT_TYPE` catalog.
 
-The artifact type catalog exists to provide stable artifact metadata and display hints. Post-run artifact discovery now runs inline: a run that exits 0 registers its output directories (`run_output_directory`, `attention_output_directory`) through `services::openfold_artifacts::register_known_openfold_artifacts`, idempotently, and `vizfold register-artifacts <id>` re-runs it. The workbench (`vizfold serve`) reads those rows, lists the registered directories, and renders each run's structure and attention images. Discovery below directory granularity — classifying individual files by artifact type — is still deferred.
+The artifact type catalog exists to provide stable artifact metadata and display hints. Post-run artifact discovery now runs inline: a run that exits 0 registers its output directories (`run_output_directory`, `attention_output_directory`) through `services::run_artifacts::register_known_run_artifacts`, idempotently, and `vizfold register-artifacts <id>` re-runs it. The workbench (`vizfold serve`) reads those rows, lists the registered directories, and renders each run's structure and attention images. Discovery below directory granularity — classifying individual files by artifact type — is still deferred.
 
 ## Architecture note: parameter and resource ownership
 
@@ -64,7 +64,7 @@ For a concrete `RUN`, the executor loads the selected model, target, invocation 
 
 Before execution the command always passes through a preflight, selected by the run's backend slug — `preflight_openfold` or `preflight_esmfold`, with unknown slugs falling back to OpenFold's. Each performs model-specific readiness checks against the planned (unwrapped) command and returns a `PreflightReport` of passed checks, warnings, or failures. If failures are reported, execution is skipped, the run is marked `failed` with the failure messages, and the report is returned.
 
-`services::openfold_execution::execute_run` coordinates this flow: create the run workspace → plan a `CommandSpec` → wrap it for the backend's environment (micromamba activation for OpenFold, the venv interpreter for ESMFold, `srun` outermost when a GPU partition is configured) → preflight the unwrapped command → `CommandRunner`. The `CommandRunner` executes the command and returns a `CommandOutput` containing the exit code, stdout, and stderr; `execute_run` moves the run through `running` → `completed`/`failed` and registers its artifacts on success.
+`services::run_execution::execute_run` coordinates this flow: create the run workspace → plan a `CommandSpec` → wrap it for the backend's environment (micromamba activation for OpenFold, the venv interpreter for ESMFold, `srun` outermost when a GPU partition is configured) → preflight the unwrapped command → `CommandRunner`. The `CommandRunner` executes the command and returns a `CommandOutput` containing the exit code, stdout, and stderr; `execute_run` moves the run through `running` → `completed`/`failed` and registers its artifacts on success.
 
 One built-in schema-driven Rust planner serves every backend — it emits OpenFold's or ESMFold's CLI from that backend's `parameter_schema_json` — and each backend contributes its own preflight. Later, the same abstractions can support DB-driven command templates, external model plugins, richer preflight checks, and additional execution targets without changing the core execution flow. Produced outputs are not stored directly in the database; they remain in external storage and are registered as `ARTIFACT` manifest entries classified by the `ARTIFACT_TYPE` catalog. `ARTIFACT_TYPE` is catalog/reference data. `ARTIFACT` is run-specific manifest data.
